refactor: validatePlaygroundToken
