+++ src/components/util/auth.ts
@@ -1,5 +1,5 @@
 import { playgroundToken } from '@/stores/tokenStore';
-import { getCrewToken } from '@api/auth';
+import { validatePlaygroundToken } from '@api/auth';
 import { playgroundLink } from '@sopt-makers/playground-common';
 
 export const ACCESS_TOKEN_KEY = 'serviceAccessToken';
@@ -11,19 +11,6 @@ export const redirectToLoginPage = () => {
 
 export const getPlaygroundToken = () => {
   return localStorage.getItem(ACCESS_TOKEN_KEY);
-};
-
-export const validatePlaygroundToken = async (playgroundToken: string) => {
-  // NOTE: playground Token으로 초기 로그인 검증
-  try {
-    await getCrewToken(playgroundToken);
-    return true;
-  } catch {
-    // TODO: 에러를 어떻게 핸들링하지?
-    alert('계정 정보를 불러오지 못했습니다. 다시 로그인 해주세요.');
-    redirectToLoginPage();
-    return false;
-  }
 };
 
 export const setAccessTokens = async () => {
